python: link all depended code into shared library
This enables us to independently use the pybookwyrm module when
developing new seekers. This wasn't possible because of a big list of
undefined symbols. Presumeably bookwyrm itself could import the module
because that binary has all needed symbols.

This is an ugly fix, but cleanup will come closer to release (I hope).

Oh, I almost forgot: fuck this mess of a language.